updated the deleted comment function

+++ src/assets/DeleteComment.jsx
@@ -1,24 +1,47 @@
-import React from "react";
+import React, { useState } from "react";
 import axios from "axios";
 import "./CSS/Comments.css";
+import UrlBase from "./UrlBase";
 
 const CommentActions = ({ comment_id, onDelete }) => {
+  const [confirmDelete, setConfirmDelete] = useState(false);
+
   const handleDelete = () => {
-    axios
-      .delete(`https://js-be-project.onrender.com/api/comments/${comment_id}`)
-      .then(() => {
-        onDelete(comment_id);
-      })
-      .catch((err) => {
-        console.log("Error deleting comment", err);
-      });
+    if (confirmDelete) {
+      axios
+        .delete(`${UrlBase}comments/${comment_id}`)
+        .then(() => {
+          onDelete(comment_id);
+        })
+        .catch((err) => {
+          console.log("Error deleting comment", err);
+        });
+    } else {
+      setConfirmDelete(true);
+    }
+  };
+
+  const handleCancelDelete = () => {
+    setConfirmDelete(false);
   };
 
   return (
     <div className="comment-actions" key={comment_id}>
-      <button className="menu__button" onClick={handleDelete}>
-        Delete
-      </button>
+      {confirmDelete ? (
+        <>
+          <p>Are you sure you want to delete this comment?</p>
+          <button className="menu__button" onClick={handleDelete}>
+            Yes
+          </button>
+          <button className="menu__button" onClick={handleCancelDelete}>
+            No
+          </button>
+        </>
+      ) : (
+        <button className="menu__button" onClick={handleDelete}>
+          Delete
+        </button>
+      )}
     </div>
   );
 };
